pyproject.toml
[tool.poetry]
name = "microchat"
version = "0.1.0"
description = "A vision-language benchmark for image-based reasoning in biomedical images."
authors = ["James Burgess, Jeffrey Nirschl, Laura Bravo"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
click = ">=8.0.1"
python-dotenv = "^1.0.0"
pandas = "^2.1.0"
loguru = "^0.7.0"
rich = "^13.7.0"
tqdm = "^4.66.2"
pillow = "^9.5.0"
openai = "^1.51.2"
matplotlib = "^3.8.4"
tenacity = "^9.0.0"
datasets = "^3.0.1"
hydra-core = "^1.3.2"

[tool.poetry.group.dspy]
optional = true
[tool.poetry.group.dspy.dependencies]
dspy = "^2.5.13"
dspy-ai = {extras = ["chromadb"], version = "^2.5.9"}

[tool.poetry.group.dvc]
optional = true
[tool.poetry.group.dvc.dependencies]
dvc = {extras = ["s3"], version = "^3.17.0"}

[tool.poetry.group.dev]
optional = true
[tool.poetry.group.dev.dependencies]
Pygments = ">=2.17.0"
autopep8 = "^1.6.0"
black = ">23.0.0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = "^4.0.1,<5.0.0"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.13.1"
mypy = ">=0.930"
nox-poetry = ">1.0.1"
Faker = ">18.0.0"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=7.0.0"
pytest-randomly = "^3.12.0"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=7.1.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}
nox = "^2022.1.7"

[tool.poetry.scripts]
microchat = "microchat.__main__:main"

[tool.black]
line-length = 88
target-version = ["py39"]
exclude = "/.github/"

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2
extend_skip = [".md", ".json"]
skip_glob = ["**/docs/**", "**/site-packages/**"]

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
addopts = [
    "-rsxX --showlocals",
]

markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "smoke: determines whether the deployed build is stable or not (select with '-m \"smoke\"')",
    "debug: marks tests as actively under debugging (deselect with '-m \"not debug\"')",
]
norecursedirs = ".git/* .dvc/* build/* data/external/* data/interim/* data/processed/* data/raw/*"

[virtualenvs]
create = true
in-project = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
